fix cleanup script(maybe?)

+++ cleanupHDWX.py
@@ -46,6 +46,8 @@ if __name__ == "__main__":
                     runPathExtension = runData["pathExtension"]
                     if "gr2a" not in productPath:
                         # Now we know where the frames for this product are located, and we can purge them!
-                        rmtree(path.join(path.join(hdwxRootPath, productPath), runPathExtension))
+                        if path.exists(path.join(hdwxRootPath, productPath, runPathExtension)):
+                            rmtree(path.join(path.join(hdwxRootPath, productPath), runPathExtension))
                         # Also remove the json data
-                        remove(runFilePath)
+                        if path.exists(runFilePath):
+                            remove(runFilePath)
